Update service arch
