+++ apps/shelve/server/api/admin/users/[userId].delete.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { deleteUser } from '~~/server/app/userService'
+import type { H3Event } from 'h3'
+import { deleteUser } from '~~/server/services/user.service'
 
 export default eventHandler(async (event: H3Event) => {
   const { user } = event.context
