+++ apps/shelve/server/api/auth/currentUser.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getUserByAuthToken } from '~~/server/app/tokenService'
+import type { H3Event } from 'h3'
+import { getUserByAuthToken } from '~~/server/services/token.service'
 
 export default eventHandler(async (event: H3Event) => {
   const authToken = getCookie(event, 'authToken') || ''
