+++ apps/shelve/server/api/github/repos.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getUserRepos } from '~~/server/app/githubService'
+import type { H3Event } from 'h3'
+import { getUserRepos } from '~~/server/services/github.service'
 
 export default defineEventHandler(async (event: H3Event) => {
   return await getUserRepos(event)
