+++ apps/shelve/server/api/project/[id]/index.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getProjectById } from '~~/server/app/projectService'
+import type { H3Event } from 'h3'
+import { getProjectById } from '~~/server/services/project.service'
 
 export default eventHandler(async (event: H3Event) => {
   const id = getRouterParam(event, 'id') as string
