+++ apps/shelve/server/api/project/[id]/team/[teamId].post.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { addTeamToProject } from '~~/server/app/projectService'
+import type { H3Event } from 'h3'
+import { addTeamToProject } from '~~/server/services/project.service'
 
 export default defineEventHandler(async (event: H3Event) => {
   const id = getRouterParam(event, 'id') as string
