+++ apps/shelve/server/api/teams/index.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getTeamByUserId } from '~~/server/app/teamsService'
+import type { H3Event } from 'h3'
+import { getTeamByUserId } from '~~/server/services/teams.service'
 
 export default eventHandler((event: H3Event) => {
   const { user } = event.context
