+++ apps/shelve/server/api/tokens/[id].delete.ts
@@ -1,5 +1,4 @@
-import { H3Event } from 'h3'
-
+import type { H3Event } from 'h3'
 
 export default defineEventHandler(async (event: H3Event) => {
   const { user } = event.context
