+++ apps/shelve/server/api/tokens/[token].get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getUserByAuthToken } from '~~/server/app/tokenService'
+import type { H3Event } from 'h3'
+import { getUserByAuthToken } from '~~/server/services/token.service'
 
 export default defineEventHandler((event: H3Event) => {
   const token = getRouterParam(event, 'token') as string
