+++ apps/shelve/server/api/tokens/index.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getTokensByUserId } from '~~/server/app/tokenService'
+import type { H3Event } from 'h3'
+import { getTokensByUserId } from '~~/server/services/token.service'
 
 export default defineEventHandler((event: H3Event) => {
   const { user } = event.context
