+++ apps/shelve/server/api/user/teammate/index.get.ts
@@ -1,5 +1,5 @@
-import { H3Event } from 'h3'
-import { getTeammatesByUserId } from '~~/server/app/teammateService'
+import type { H3Event } from 'h3'
+import { getTeammatesByUserId } from '~~/server/services/teammate.service'
 
 export default eventHandler((event: H3Event) => {
   const { user } = event.context
