+++ apps/shelve/server/api/variable/[id]/[env].get.ts
@@ -1,6 +1,6 @@
-import { Environment } from '@shelve/types'
-import { H3Event } from 'h3'
-import { getVariablesByProjectId } from '~~/server/app/variableService'
+import type { Environment } from '@shelve/types'
+import type { H3Event } from 'h3'
+import { getVariablesByProjectId } from '~~/server/services/variable.service'
 
 export default eventHandler(async (event: H3Event) => {
   const id = getRouterParam(event, 'id') as string
